+++ internal/ical/import.go
@@ -8,7 +8,6 @@ import (
 	"fmt"
 	"io"
 	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -222,216 +221,6 @@ func importFile(r io.Reader, remote bool) (ImportResult, error) {
 	return result, nil
 }
 
-// todoSpanStorable reports whether the DTSTART plus the DURATION lands on
-// a time the database can hold. See timeutil.Storable for that rule. An
-// unparseable start gives no end to test, so the function accepts it. The
-// todo service applies the same rule in validateTiming.
-func todoSpanStorable(startDate, dur string) bool {
-	start := timeutil.ParseDate(startDate)
-	if start.IsZero() {
-		return true
-	}
-	return timeutil.Storable(duration.Add(start, dur))
-}
-
-func todoFromVTodo(comp *ical.Component) (todo.Todo, []string, error) {
-	props := comp.Props
-
-	uid := propText(props, ical.PropUID)
-	if uid == "" {
-		return todo.Todo{}, nil, fmt.Errorf("missing UID")
-	}
-
-	summary := propText(props, ical.PropSummary)
-	description := propText(props, ical.PropDescription)
-	location := propText(props, ical.PropLocation)
-
-	var todoWarnings []string
-	dueDate, w := parseDateProp(props, "todo", ical.PropDue, uid)
-	if w != "" {
-		todoWarnings = append(todoWarnings, w)
-	}
-	startDate, w := parseDateProp(props, "todo", ical.PropDateTimeStart, uid)
-	if w != "" {
-		todoWarnings = append(todoWarnings, w)
-	}
-
-	// Screen the DURATION on the same four rules the todo service
-	// applies. The value must be a positive span (RFC 5545 §3.8.2.5).
-	// DUE and DURATION are mutually exclusive, and a DURATION needs a
-	// DTSTART (§3.6.2). The end must also land on a time the database
-	// can hold.
-	//
-	// Drop the DURATION and warn when a rule fails. A stored bad value
-	// would re-export verbatim and would poison a RELATED=END alarm
-	// anchor. A rejected shape would fail the whole calendar pull on
-	// every run, because the sync engine stops at the first resource
-	// error.
-	var durationValue string
-	if prop := props.Get(ical.PropDuration); prop != nil {
-		switch {
-		case duration.ValidateSpan(prop.Value) != nil:
-			todoWarnings = append(todoWarnings, fmt.Sprintf(
-				"todo %q: unusable DURATION %q; dropped", uid, prop.Value))
-		case dueDate != "":
-			todoWarnings = append(todoWarnings, fmt.Sprintf(
-				"todo %q: DUE and DURATION are mutually exclusive; DURATION %q dropped", uid, prop.Value))
-		case startDate == "":
-			todoWarnings = append(todoWarnings, fmt.Sprintf(
-				"todo %q: DURATION %q needs a DTSTART; dropped", uid, prop.Value))
-		case !todoSpanStorable(startDate, prop.Value):
-			todoWarnings = append(todoWarnings, fmt.Sprintf(
-				"todo %q: DURATION %q ends past year %d; dropped", uid, prop.Value, timeutil.MaxStorableYear))
-		default:
-			durationValue = prop.Value
-		}
-	}
-
-	var completedAt string
-	if prop := props.Get(ical.PropCompleted); prop != nil {
-		if t, err := prop.DateTime(nil); err == nil && !t.IsZero() {
-			completedAt = t.UTC().Format(time.RFC3339)
-		}
-	}
-
-	var percentComplete int64
-	if prop := props.Get(ical.PropPercentComplete); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			percentComplete = v
-		}
-	}
-
-	status := propTextOr(props, ical.PropStatus, "NEEDS-ACTION")
-	class := propTextOr(props, ical.PropClass, "PUBLIC")
-
-	var priority int64
-	if prop := props.Get(ical.PropPriority); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			priority = v
-		}
-	}
-
-	var sequence int64
-	if prop := props.Get("SEQUENCE"); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			sequence = v
-		}
-	}
-
-	url := propText(props, ical.PropURL)
-
-	var timezone string
-	var todoFloating bool
-	if prop := props.Get(ical.PropDateTimeStart); prop != nil {
-		if tzid := prop.Params.Get(ical.ParamTimezoneID); tzid != "" {
-			timezone = tzid
-		} else if len(prop.Value) > 8 && !strings.HasSuffix(prop.Value, "Z") {
-			todoFloating = true
-		}
-	}
-	if timezone == "" && !todoFloating {
-		if prop := props.Get(ical.PropDue); prop != nil {
-			if tzid := prop.Params.Get(ical.ParamTimezoneID); tzid != "" {
-				timezone = tzid
-			} else if len(prop.Value) > 8 && !strings.HasSuffix(prop.Value, "Z") {
-				todoFloating = true
-			}
-		}
-	}
-
-	categories := parseCategoriesFromProps(props)
-	exdates := parseDateListFromProps(props, ical.PropExceptionDates, dtstartZone(props))
-	rdates := parseDateListFromProps(props, ical.PropRecurrenceDates, dtstartZone(props))
-	var rrule string
-	if prop := props.Get(ical.PropRecurrenceRule); prop != nil {
-		rrule = prop.Value
-	}
-
-	var geo string
-	if prop := props.Get(ical.PropGeo); prop != nil {
-		geo = prop.Value
-	}
-
-	recurrenceID, err := parseRecurrenceID(props)
-	if err != nil {
-		return todo.Todo{}, nil, err
-	}
-
-	var dtstamp string
-	if prop := props.Get(ical.PropDateTimeStamp); prop != nil {
-		if t, err := prop.DateTime(nil); err == nil && !t.IsZero() {
-			dtstamp = t.UTC().Format(time.RFC3339)
-		}
-	}
-
-	// VALARM children
-	var alarms []model.Alarm
-	var alarmWarnings []string
-	for _, child := range comp.Children {
-		if child.Name != ical.CompAlarm {
-			continue
-		}
-		alarm, w := parseAlarm(child)
-		if w != "" {
-			// Name the owning record: the dropped alarm leaves no other trace.
-			alarmWarnings = append(alarmWarnings, fmt.Sprintf("todo %q: %s", uid, w))
-		}
-		if alarm.TriggerValue != "" {
-			alarms = append(alarms, alarm)
-		}
-	}
-
-	// ATTENDEE + ORGANIZER
-	attendees, attendeeWarnings := parseAttendeesFromProps(props, model.TaskAttendee)
-	for _, w := range attendeeWarnings {
-		// Name the owning record: the clamp leaves no other trace.
-		todoWarnings = append(todoWarnings, fmt.Sprintf("todo %q: %s", uid, w))
-	}
-
-	// ATTACH, COMMENT, CONTACT, RELATED-TO
-	attachments, err := parseAttachmentsFromProps(props)
-	if err != nil {
-		return todo.Todo{}, nil, err
-	}
-	comments := parseCommentsFromProps(props)
-	contacts := parseContactsFromProps(props)
-	resources := parseResourcesFromProps(props)
-	relations := parseRelationsFromProps(props)
-
-	return todo.Todo{
-		UID:             uid,
-		Summary:         summary,
-		Description:     description,
-		Location:        location,
-		DueDate:         dueDate,
-		StartDate:       startDate,
-		Duration:        durationValue,
-		CompletedAt:     completedAt,
-		PercentComplete: percentComplete,
-		Status:          strings.ToUpper(status),
-		Priority:        priority,
-		Class:           strings.ToUpper(class),
-		URL:             url,
-		Categories:      categories,
-		RecurrenceRule:  rrule,
-		Timezone:        floatingOrTZ(todoFloating, timezone),
-		Sequence:        sequence,
-		ExDates:         exdates,
-		RDates:          rdates,
-		RecurrenceID:    recurrenceID,
-		Geo:             geo,
-		DtStamp:         dtstamp,
-		Alarms:          alarms,
-		Attendees:       attendees,
-		Attachments:     attachments,
-		Comments:        comments,
-		Contacts:        contacts,
-		Resources:       resources,
-		Relations:       relations,
-		XProperties:     extractXPropertiesWithSet(props, handledTodoProps),
-	}, append(alarmWarnings, todoWarnings...), nil
-}
-
 // parseDateProp formats a component's date/date-time property for storage.
 // It returns a warning rather than a silent discard of an unparseable value.
 // A dropped DUE or DTSTART is invisible data loss. The record disappears from
@@ -452,547 +241,6 @@ func parseDateProp(props ical.Props, kind, name, uid string) (string, string) {
 		return t.Format("2006-01-02"), ""
 	}
 	return t.UTC().Format(time.RFC3339), ""
-}
-
-func eventFromVEvent(ve ical.Event, remote bool) (event.Event, []string, error) {
-	uid, err := ve.Props.Text(ical.PropUID)
-	if err != nil || uid == "" {
-		return event.Event{}, nil, fmt.Errorf("missing UID")
-	}
-
-	summary, _ := ve.Props.Text(ical.PropSummary)
-	description, _ := ve.Props.Text(ical.PropDescription)
-	location, _ := ve.Props.Text(ical.PropLocation)
-
-	// Timezone from DTSTART param
-	var timezone string
-	var floating bool
-	if prop := ve.Props.Get(ical.PropDateTimeStart); prop != nil {
-		tzid := prop.Params.Get(ical.ParamTimezoneID)
-		if tzid != "" {
-			timezone = tzid
-		} else if !strings.EqualFold(prop.Params.Get("VALUE"), "DATE") &&
-			!strings.HasSuffix(prop.Value, "Z") {
-			// No TZID, not all-day, no Z suffix → floating time.
-			floating = true
-		}
-	}
-
-	startTime, err := ve.Props.DateTime(ical.PropDateTimeStart, nil)
-	if err != nil {
-		return event.Event{}, nil, fmt.Errorf("parse DTSTART: %w", err)
-	}
-
-	var endTime time.Time
-	var durationValue string
-	var dtendWarnings []string
-	explicitEnd := false
-	badDTEND := false
-	var badDTENDRaw *ical.Prop
-	if prop := ve.Props.Get(ical.PropDateTimeEnd); prop != nil {
-		var err error
-		endTime, err = prop.DateTime(nil)
-		if err != nil {
-			// Malformed DTEND (go-ical stores the raw value without
-			// validating). Props.Get above guarantees the property exists, so
-			// the only error DateTime can return here is a parse failure.
-			//
-			// Discard it and fall through as if DTEND were absent: a valid
-			// DURATION on the same component is still the better answer, and
-			// leaving explicitEnd false keeps the all-day branch below free to
-			// apply the RFC 5545 implicit one-day span. Forcing an explicit end
-			// here instead would collapse an all-day event to zero duration —
-			// the exact outcome this fallback exists to prevent.
-			endTime = time.Time{}
-			badDTEND = true
-			badDTENDRaw = prop
-			dtendWarnings = append(dtendWarnings, fmt.Sprintf(
-				"event %q: unparseable DTEND %q; ignored, falling back to DURATION or the default span", uid, prop.Value))
-		}
-	}
-	if endTime.IsZero() {
-		prop := ve.Props.Get(ical.PropDuration)
-		spanOK := false
-		if prop != nil && duration.ValidateSpan(prop.Value) == nil {
-			// The end must land on a time the database can hold. See
-			// timeutil.Storable for that rule.
-			if end := addDuration(startTime, prop.Value); timeutil.Storable(end) {
-				durationValue = prop.Value
-				endTime = end
-				explicitEnd = true
-				spanOK = true
-			}
-		}
-		if spanOK {
-			// The DURATION above is the span.
-		} else if prop != nil && duration.Validate(prop.Value) == nil &&
-			addDuration(startTime, prop.Value).Equal(startTime) {
-			// An exact zero span (DURATION:PT0S). RFC 5545 §3.6.1 gives
-			// the same meaning to a VEVENT with no DTEND and no
-			// DURATION, so store the equivalent zero-length event. Drop
-			// the value; DTEND=DTSTART carries the semantics on export.
-			endTime = startTime
-			explicitEnd = true
-		} else if prop != nil {
-			// Malformed, negative, or unstorable DURATION (go-ical
-			// stores the raw value with no validation). Fall back to a
-			// 1h span. Warn the user. Drop the bad value, so it does
-			// not persist and does not re-export.
-			endTime = startTime.Add(time.Hour)
-			dtendWarnings = append(dtendWarnings, fmt.Sprintf(
-				"event %q: unusable DURATION %q; dropped, the event gets a 1h span", uid, prop.Value))
-		} else if badDTEND {
-			// DTEND was present but unusable and there is no DURATION to fall
-			// back on. A timed event gets the same 1h default as a malformed
-			// DURATION rather than collapsing to zero duration; an all-day one
-			// falls through to the implicit one-day span below, since
-			// explicitEnd is still false.
-			endTime = startTime.Add(time.Hour)
-		} else {
-			// RFC 5545 §3.6.1: a VEVENT with DTSTART as DATE-TIME and no
-			// DTEND/DURATION is instantaneous (zero duration). The all-day
-			// case (VALUE=DATE) is handled below to apply the implicit one-day
-			// span instead.
-			endTime = startTime
-		}
-	} else {
-		explicitEnd = true
-	}
-
-	allDay := false
-	if prop := ve.Props.Get(ical.PropDateTimeStart); prop != nil {
-		if strings.EqualFold(prop.Params.Get("VALUE"), "DATE") {
-			allDay = true
-			// VALUE=DATE represents a calendar date, not a specific UTC instant.
-			// Store as midnight UTC so the stored instant is independent of the
-			// importing host's timezone (AGENTS.md: "All database times are
-			// RFC 3339 strings in UTC"; "All-day events have time component
-			// 00:00:00"). Using time.Local here would store a host-dependent
-			// instant — e.g. under TZ=UTC+12 midnight local is 12:00Z the day
-			// before, corrupting the calendar date and recurrence occurrences.
-			startTime = time.Date(startTime.Year(), startTime.Month(), startTime.Day(), 0, 0, 0, 0, time.UTC)
-			if explicitEnd {
-				endTime = time.Date(endTime.Year(), endTime.Month(), endTime.Day(), 0, 0, 0, 0, time.UTC)
-			} else {
-				// RFC 5545 §3.6.1: an all-day VEVENT with no DTEND/DURATION has
-				// an implicit duration of one day.
-				endTime = startTime.AddDate(0, 0, 1)
-			}
-		}
-	}
-
-	var rrule string
-	if prop := ve.Props.Get(ical.PropRecurrenceRule); prop != nil {
-		rrule = prop.Value
-	}
-
-	// RFC 5545 properties
-	status := textOrDefault(ve, ical.PropStatus, "CONFIRMED")
-	transp := textOrDefault(ve, ical.PropTransparency, "OPAQUE")
-	class := textOrDefault(ve, ical.PropClass, "PUBLIC")
-
-	var sequence int64
-	if prop := ve.Props.Get("SEQUENCE"); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			sequence = v
-		}
-	}
-
-	var priority int64
-	if prop := ve.Props.Get(ical.PropPriority); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			priority = v
-		}
-	}
-
-	var url string
-	if prop := ve.Props.Get(ical.PropURL); prop != nil {
-		url = prop.Value
-	}
-
-	var conferenceURI string
-	if prop := ve.Props.Get("CONFERENCE"); prop != nil {
-		conferenceURI = prop.Value
-	}
-
-	var geo string
-	if prop := ve.Props.Get(ical.PropGeo); prop != nil {
-		geo = prop.Value
-	}
-
-	categories := parseCategories(ve)
-	exdates := parseDateListFromProps(ve.Props, ical.PropExceptionDates, dtstartZone(ve.Props))
-	rdates := parseDateListFromProps(ve.Props, ical.PropRecurrenceDates, dtstartZone(ve.Props))
-
-	recurrenceID, err := parseRecurrenceID(ve.Props)
-	if err != nil {
-		return event.Event{}, nil, err
-	}
-
-	var dtstamp string
-	if prop := ve.Props.Get(ical.PropDateTimeStamp); prop != nil {
-		if t, err := prop.DateTime(nil); err == nil && !t.IsZero() {
-			dtstamp = t.UTC().Format(time.RFC3339)
-		}
-	}
-
-	// VALARM children
-	var alarms []model.Alarm
-	var childWarnings []string
-	for _, child := range ve.Children {
-		if child.Name != ical.CompAlarm {
-			continue
-		}
-		alarm, w := parseAlarm(child)
-		if w != "" {
-			// Name the owning record: the dropped alarm leaves no other trace.
-			childWarnings = append(childWarnings, fmt.Sprintf("event %q: %s", uid, w))
-		}
-		if alarm.TriggerValue != "" {
-			alarms = append(alarms, alarm)
-		}
-	}
-
-	// ATTENDEE + ORGANIZER
-	attendees, attendeeWarnings := parseAttendees(ve)
-	for _, w := range attendeeWarnings {
-		// Name the owning record: the clamp leaves no other trace.
-		childWarnings = append(childWarnings, fmt.Sprintf("event %q: %s", uid, w))
-	}
-
-	// ATTACH, COMMENT, RELATED-TO
-	attachments, err := parseAttachmentsFromProps(ve.Props)
-	if err != nil {
-		return event.Event{}, nil, err
-	}
-	comments := parseCommentsFromProps(ve.Props)
-	contacts := parseContactsFromProps(ve.Props)
-	resources := parseResourcesFromProps(ve.Props)
-	relations := parseRelationsFromProps(ve.Props)
-
-	xprops := extractXPropertiesWithSet(ve.Props, handledEventProps)
-	if remote && badDTENDRaw != nil {
-		// The server's DTEND failed to parse here but parses on the
-		// server (a non-IANA TZID is the usual cause). Store the raw
-		// property, so export returns the server value instead of the
-		// fabricated span (issue #567).
-		params := "{}"
-		if len(badDTENDRaw.Params) > 0 {
-			if b, err := json.Marshal(badDTENDRaw.Params); err == nil {
-				params = string(b)
-			}
-		}
-		xprops = append(xprops, model.XProperty{
-			Name:   xpropOriginalDTEND,
-			Value:  badDTENDRaw.Value,
-			Params: params,
-		})
-	}
-	return event.Event{
-		UID:            uid,
-		Title:          summary,
-		Description:    description,
-		Location:       location,
-		StartTime:      startTime.UTC(),
-		EndTime:        endTime.UTC(),
-		AllDay:         allDay,
-		RecurrenceRule: rrule,
-		Timezone:       floatingOrTZ(floating, timezone),
-		Status:         strings.ToUpper(status),
-		Transp:         strings.ToUpper(transp),
-		Sequence:       sequence,
-		Priority:       priority,
-		Class:          strings.ToUpper(class),
-		URL:            url,
-		ConferenceURI:  conferenceURI,
-		Categories:     categories,
-		ExDates:        exdates,
-		RDates:         rdates,
-		RecurrenceID:   recurrenceID,
-		Geo:            geo,
-		DurationValue:  durationValue,
-		DtStamp:        dtstamp,
-		Alarms:         alarms,
-		Attendees:      attendees,
-		Attachments:    attachments,
-		Comments:       comments,
-		Contacts:       contacts,
-		Resources:      resources,
-		Relations:      relations,
-		XProperties:    xprops,
-	}, append(childWarnings, dtendWarnings...), nil
-}
-
-func textOrDefault(ve ical.Event, prop, def string) string {
-	if v, err := ve.Props.Text(prop); err == nil && v != "" {
-		return v
-	}
-	return def
-}
-
-func parseCategories(ve ical.Event) string {
-	var cats []string
-	for _, prop := range ve.Props.Values(ical.PropCategories) {
-		// TextList splits on unescaped commas and unescapes each value,
-		// handling both RFC-correct "CATEGORIES:a,b" and legacy
-		// escaped "CATEGORIES:a\,b" inputs.
-		if list, err := prop.TextList(); err == nil {
-			cats = append(cats, list...)
-		}
-	}
-	// Join with comma-escaping so a category value that itself contains a
-	// comma (e.g. "Foo, Bar") stays a single value through the in-memory
-	// comma-joined representation instead of fragmenting on round-trip.
-	return timeutil.JoinCategoryList(cats)
-}
-
-// parseAlarm extracts a model.Alarm from a VALARM component.
-// The second return value is a warning string (empty if no issues). A VALARM
-// with several problems reports all of them. Each one changes what the alarm
-// does in silence. The user needs to see every reason.
-func parseAlarm(comp *ical.Component) (model.Alarm, string) {
-	alarm := model.Alarm{Action: model.DefaultAlarmAction, Related: model.DefaultAlarmRelated}
-	var warns []string
-
-	if prop := comp.Props.Get(ical.PropAction); prop != nil {
-		// The parser preserves an action outside model.FireableAlarmAction
-		// (issue #579; that predicate's doc comment holds the shared
-		// rationale). A drop makes every later push delete the VALARM of
-		// the other client. The parser stores the trimmed original case,
-		// so the foreign VALARM round-trips verbatim. A fireable action
-		// stays canonical uppercase.
-		raw := strings.TrimSpace(prop.Value)
-		switch up := strings.ToUpper(raw); {
-		case up == "":
-			// An empty value keeps the default action. The service
-			// write boundary (model.PrepareAlarmsForWrite) fills the
-			// same default, and a reminder must not vanish over an
-			// empty value.
-		case model.FireableAlarmAction(up):
-			alarm.Action = up
-		case model.ValidAlarmActionToken(raw):
-			alarm.Action = raw
-		default:
-			// A malformed action cannot round-trip: export would emit
-			// an invalid ACTION line, and a strict server rejects the
-			// whole resource with 400. Drop the alarm and warn.
-			warns = append(warns, fmt.Sprintf("VALARM ACTION %q: not an RFC 5545 token; alarm dropped", raw))
-			return model.Alarm{}, strings.Join(warns, "; ")
-		}
-	}
-	if prop := comp.Props.Get(ical.PropTrigger); prop != nil {
-		tv := prop.Value
-		tzid := prop.Params.Get(ical.ParamTimezoneID)
-		// Validate the trigger: one arm per value SHAPE, each format parsed
-		// exactly once, with TZID consulted only where it matters (the
-		// compact-floating form). Keep the accepted set in lockstep with
-		// model.ParseAbsoluteTime — that function defines what a stored
-		// absolute trigger MEANS, so a form accepted here but not there
-		// would store alarms the fire path cannot read.
-		isDuration := duration.Validate(tv) == nil
-		valid := isDuration
-		if !valid {
-			if _, err := time.Parse("20060102T150405Z", tv); err == nil {
-				// Already compact UTC — the canonical stored form.
-				valid = true
-			} else if t, err := time.Parse(time.RFC3339, tv); err == nil {
-				// RFC 3339 carries its own offset (a TZID param, if present,
-				// is redundant): normalize to compact UTC so a stored
-				// absolute trigger has one encoding. Safe because the offset
-				// is explicit — this is NOT the floating-value normalization
-				// that was reverted (issue #572).
-				tv = t.UTC().Format("20060102T150405Z")
-				valid = true
-			} else if t, err := time.Parse("20060102T150405", tv); err == nil {
-				// Compact floating: resolve through the TZID when it loads;
-				// otherwise keep the raw floating value, to be resolved
-				// against the record's timezone at fire time.
-				if tzid != "" {
-					if loc, lerr := time.LoadLocation(tzid); lerr == nil {
-						t = time.Date(t.Year(), t.Month(), t.Day(),
-							t.Hour(), t.Minute(), t.Second(), 0, loc)
-						tv = t.UTC().Format("20060102T150405Z")
-					} else {
-						warns = append(warns, fmt.Sprintf("VALARM TRIGGER TZID=%s: unknown timezone, treating as floating", tzid))
-					}
-				}
-				valid = true
-			}
-		}
-		if valid {
-			alarm.TriggerValue = tv
-		} else {
-			// Unparseable (or empty) TRIGGER: leave TriggerValue empty so
-			// the caller's `TriggerValue != ""` gate drops the alarm, and
-			// warn.
-			//
-			// Preserving the raw value here looks like it protects round-trip
-			// fidelity, but it cannot. The value is not expressible as valid
-			// iCal, so the next push either emits a VALARM strict servers
-			// reject with 400 — wedging the whole resource — or omits it and
-			// deletes the alarm from the server anyway. Preserving only moves
-			// that loss from "announced at import" to "silent at the next
-			// push", and in the meantime stores an alarm every trigger-time
-			// helper refuses while the CLI and TUI display it as an armed
-			// reminder the alarm editor will not even open. Losing it loudly
-			// is the honest trade.
-			warns = append(warns, fmt.Sprintf("VALARM TRIGGER: unparseable value %q; alarm dropped (it could never fire)", tv))
-		}
-		// RELATED only means something on a duration trigger (it picks the
-		// anchor the offset applies to). RFC 5545 §3.8.6.3's trigabs
-		// production forbids RELATED on an absolute trigger, so export can
-		// never emit it — a stored one would be junk that never round-trips
-		// (push+pull resets it to START) and that silently resurfaces if the
-		// user later switches the trigger to a duration. Keep the default
-		// "START" for absolute triggers.
-		if rel := strings.TrimSpace(prop.Params.Get("RELATED")); rel != "" && isDuration {
-			// Same failure class as an unsupported ACTION (issue #575,
-			// see model.ValidAlarmRelated).
-			if up := strings.ToUpper(rel); model.ValidAlarmRelated(up) {
-				alarm.Related = up
-			} else {
-				warns = append(warns, fmt.Sprintf("VALARM TRIGGER RELATED=%q: unsupported value, using %s", rel, model.DefaultAlarmRelated))
-			}
-		}
-	} else {
-		// RFC 5545 requires TRIGGER. Without it the caller's gate drops
-		// the alarm, and the drop must not be silent.
-		warns = append(warns, "VALARM TRIGGER: missing; alarm dropped (it could never fire)")
-	}
-	if prop := comp.Props.Get(ical.PropDescription); prop != nil {
-		alarm.Description = prop.Value
-	}
-	if prop := comp.Props.Get(ical.PropSummary); prop != nil {
-		alarm.Summary = prop.Value
-	}
-	repeatZero := false
-	clampedFrom := 0
-	if prop := comp.Props.Get("REPEAT"); prop != nil {
-		// Clamp: REPEAT expands into per-trigger state in the check loop,
-		// so an absurd imported value must not hang or OOM every check.
-		// The clamp warning waits until after the pairing rule below.
-		// Without the wait, one report could claim the clamp survived
-		// and the repeat was dropped at the same time.
-		v, err := strconv.Atoi(strings.TrimSpace(prop.Value))
-		switch {
-		case err != nil || v < 0:
-			warns = append(warns, fmt.Sprintf("VALARM REPEAT: invalid value %q; ignored", prop.Value))
-		case v > model.MaxAlarmRepeat && !model.FireableAlarmAction(alarm.Action):
-			// A preserved sync-only alarm never expands into trigger
-			// state, so the clamp guards nothing. A clamp here would
-			// rewrite the VALARM of another client on the next push.
-			alarm.Repeat = v
-		case v > model.MaxAlarmRepeat:
-			alarm.Repeat = model.MaxAlarmRepeat
-			clampedFrom = v
-		case v > 0:
-			alarm.Repeat = v
-		default:
-			repeatZero = true // a valid "no repeats", not a defect
-		}
-	}
-	if prop := comp.Props.Get(ical.PropDuration); prop != nil {
-		// A malformed DURATION pushes verbatim, and a strict CalDAV
-		// server rejects the whole resource with 400. See
-		// model.ValidAlarmDuration for the value rule.
-		if v := strings.TrimSpace(prop.Value); model.ValidAlarmDuration(v) {
-			alarm.Duration = v
-		} else if alarm.Repeat > 0 {
-			// One defect, one warning: the REPEAT is unusable without
-			// its interval. Name both drops here, so the pairing check
-			// below finds a complete pair and adds no second warning.
-			alarm.Repeat = 0
-			warns = append(warns, fmt.Sprintf("VALARM DURATION: invalid value %q; dropped with its REPEAT", prop.Value))
-		} else {
-			warns = append(warns, fmt.Sprintf("VALARM DURATION: invalid value %q; dropped", prop.Value))
-		}
-	}
-	// RFC 5545 §3.8.6.3 requires REPEAT and DURATION together. An unpaired
-	// value cannot round-trip: buildValarm omits it, and the next pull then
-	// deletes and recreates the alarm row, which loses the alarm state.
-	// Clear the incomplete pair and name the dropped side. An explicit
-	// REPEAT:0 with a DURATION is legal iCal. The row cannot store
-	// "repeats disabled" apart from "REPEAT absent", so that pair cannot
-	// round-trip either. It gets its own accurate message.
-	switch {
-	case repeatZero && alarm.Duration != "":
-		alarm.Duration = ""
-		warns = append(warns, "VALARM REPEAT:0: repeats disabled; DURATION dropped")
-	case !alarm.RepeatPaired():
-		dropped := "REPEAT"
-		if alarm.Duration != "" {
-			dropped = "DURATION"
-		}
-		alarm.Repeat = 0
-		alarm.Duration = ""
-		warns = append(warns, fmt.Sprintf("VALARM: REPEAT and DURATION must appear together; %s dropped", dropped))
-	}
-	if clampedFrom > 0 && alarm.Repeat > 0 {
-		warns = append(warns, fmt.Sprintf("VALARM REPEAT: value %d clamped to %d", clampedFrom, model.MaxAlarmRepeat))
-	}
-
-	// UID (RFC 9074)
-	if prop := comp.Props.Get(ical.PropUID); prop != nil {
-		uid := strings.TrimSpace(prop.Value)
-		if len(uid) > 0 && len(uid) <= 255 && !strings.ContainsRune(uid, 0) {
-			alarm.UID = uid
-		}
-	}
-
-	// ACKNOWLEDGED (RFC 9074) — preserved for round-trip fidelity only.
-	if prop := comp.Props.Get("ACKNOWLEDGED"); prop != nil {
-		v := strings.TrimSpace(prop.Value)
-		if model.ValidateAcknowledged(v) && v != "" {
-			alarm.Acknowledged = v
-		}
-	}
-
-	// ATTACH (sound for AUDIO alarms): either a URI or an inline BASE64 blob.
-	if prop := comp.Props.Get(ical.PropAttach); prop != nil {
-		if prop.Params.Get("ENCODING") == "BASE64" {
-			if data, err := decodeInlineAttachment(prop.Value); err != nil {
-				warns = append(warns, fmt.Sprintf("VALARM ATTACH: %v", err))
-			} else {
-				alarm.AttachBinary = data
-				alarm.AttachFmtType = prop.Params.Get("FMTTYPE")
-			}
-		} else {
-			alarm.AttachURI = prop.Value
-			alarm.AttachFmtType = prop.Params.Get("FMTTYPE")
-		}
-	}
-
-	// ATTENDEE children (for EMAIL alarms)
-	for _, prop := range comp.Props.Values(ical.PropAttendee) {
-		alarm.Attendees = append(alarm.Attendees, model.AlarmAttendee{
-			Email: stripMailto(prop.Value),
-			Name:  prop.Params.Get(ical.ParamCommonName),
-		})
-	}
-
-	// Every property parseAlarm does not read — preserved for round-trip
-	// fidelity only. The set covers an X- extension and an IANA property
-	// alike, for example the RFC 9074 PROXIMITY. A drop would rewrite the
-	// VALARM of another client without it on the next push (issue #586).
-	// Normalize to a non-nil slice: for ReplaceAlarms, non-nil means "this
-	// is the complete X-prop set" (empty = remote cleared them), while nil
-	// means "caller has no X-prop knowledge, keep stored rows".
-	alarm.XProperties = extractXPropertiesWithSet(comp.Props, handledAlarmProps)
-	if alarm.XProperties == nil {
-		alarm.XProperties = []model.XProperty{}
-	}
-
-	// A preserved sync-only action needs a diagnostic: a legacy
-	// ACTION:PROCEDURE or a typo would otherwise look like an armed
-	// reminder. Warn only when the alarm survives the TRIGGER gate. The
-	// report must not say "preserved" and "dropped" for one alarm. Sync
-	// re-imports a resource only when it changes, so the warning does
-	// not repeat on every pull.
-	if alarm.TriggerValue != "" && !model.FireableAlarmAction(alarm.Action) {
-		warns = append(warns, fmt.Sprintf("VALARM ACTION %q: preserved for sync; it never fires locally", alarm.Action))
-	}
-
-	return alarm, strings.Join(warns, "; ")
 }
 
 // parseAttendees reads the attendees of a VEVENT. The second return
@@ -1396,175 +644,6 @@ func parseRelationsFromProps(props ical.Props) []model.Relation {
 		}
 	}
 	return out
-}
-
-func journalFromVJournal(comp *ical.Component) (journal.Journal, []string, error) {
-	props := comp.Props
-
-	uid := propText(props, ical.PropUID)
-	if uid == "" {
-		return journal.Journal{}, nil, fmt.Errorf("missing UID")
-	}
-
-	summary := propText(props, ical.PropSummary)
-
-	// VJOURNAL can have multiple DESCRIPTION properties (RFC 5545). Keep each
-	// one in the Descriptions slice for round-trip fidelity, and join them into
-	// the single DB-backed Description field used for display and search.
-	var descriptions []string
-	for _, prop := range props.Values(ical.PropDescription) {
-		text, err := prop.Text()
-		if err != nil {
-			text = prop.Value
-		}
-		if text != "" {
-			descriptions = append(descriptions, text)
-		}
-	}
-	description := strings.Join(descriptions, "\n\n")
-
-	startDate, startWarn := parseDateProp(props, "journal", ical.PropDateTimeStart, uid)
-	var journalWarnings []string
-	if startWarn != "" {
-		journalWarnings = append(journalWarnings, startWarn)
-	}
-
-	status := propTextOr(props, ical.PropStatus, "FINAL")
-	class := propTextOr(props, ical.PropClass, "PUBLIC")
-
-	var sequence int64
-	if prop := props.Get("SEQUENCE"); prop != nil {
-		if v, err := strconv.ParseInt(prop.Value, 10, 64); err == nil {
-			sequence = v
-		}
-	}
-
-	url := propText(props, ical.PropURL)
-
-	var timezone string
-	var journalFloating bool
-	if prop := props.Get(ical.PropDateTimeStart); prop != nil {
-		if tzid := prop.Params.Get(ical.ParamTimezoneID); tzid != "" {
-			timezone = tzid
-		} else if len(prop.Value) > 8 && !strings.HasSuffix(prop.Value, "Z") {
-			journalFloating = true
-		}
-	}
-
-	categories := parseCategoriesFromProps(props)
-	exdates := parseDateListFromProps(props, ical.PropExceptionDates, dtstartZone(props))
-	rdates := parseDateListFromProps(props, ical.PropRecurrenceDates, dtstartZone(props))
-	var rrule string
-	if prop := props.Get(ical.PropRecurrenceRule); prop != nil {
-		rrule = prop.Value
-	}
-
-	recurrenceID, err := parseRecurrenceID(props)
-	if err != nil {
-		return journal.Journal{}, nil, err
-	}
-
-	var dtstamp string
-	if prop := props.Get(ical.PropDateTimeStamp); prop != nil {
-		if t, err := prop.DateTime(nil); err == nil && !t.IsZero() {
-			dtstamp = t.UTC().Format(time.RFC3339)
-		}
-	}
-
-	// ATTENDEE + ORGANIZER
-	attendees, attendeeWarnings := parseAttendeesFromProps(props, model.TaskAttendee)
-	for _, w := range attendeeWarnings {
-		// Name the owning record: the clamp leaves no other trace.
-		journalWarnings = append(journalWarnings, fmt.Sprintf("journal %q: %s", uid, w))
-	}
-
-	// ATTACH, COMMENT, CONTACT, RELATED-TO
-	attachments, err := parseAttachmentsFromProps(props)
-	if err != nil {
-		return journal.Journal{}, nil, err
-	}
-	comments := parseCommentsFromProps(props)
-	contacts := parseContactsFromProps(props)
-	relations := parseRelationsFromProps(props)
-
-	return journal.Journal{
-		UID:            uid,
-		Summary:        summary,
-		Description:    description,
-		Descriptions:   descriptions,
-		StartDate:      startDate,
-		Status:         strings.ToUpper(status),
-		Class:          strings.ToUpper(class),
-		URL:            url,
-		Categories:     categories,
-		RecurrenceRule: rrule,
-		Timezone:       floatingOrTZ(journalFloating, timezone),
-		Sequence:       sequence,
-		ExDates:        exdates,
-		RDates:         rdates,
-		RecurrenceID:   recurrenceID,
-		DtStamp:        dtstamp,
-		Attendees:      attendees,
-		Attachments:    attachments,
-		Comments:       comments,
-		Contacts:       contacts,
-		XProperties:    extractXPropertiesWithSet(props, handledJournalProps),
-		Relations:      relations,
-	}, journalWarnings, nil
-}
-
-// handledAlarmProps is the set of property names parseAlarm reads into
-// the alarm model. parseAlarm preserves every other property, so a
-// foreign VALARM round-trips whole (issue #586).
-var handledAlarmProps = map[string]bool{
-	ical.PropAction: true, ical.PropTrigger: true, ical.PropDescription: true,
-	ical.PropSummary: true, "REPEAT": true, ical.PropDuration: true,
-	ical.PropUID: true, "ACKNOWLEDGED": true, ical.PropAttach: true,
-	ical.PropAttendee: true,
-}
-
-// handledEventProps is the set of property names explicitly parsed by eventFromVEvent.
-var handledEventProps = map[string]bool{
-	ical.PropUID: true, ical.PropSummary: true, ical.PropDescription: true,
-	ical.PropLocation: true, ical.PropDateTimeStart: true, ical.PropDateTimeEnd: true,
-	ical.PropDuration: true, ical.PropRecurrenceRule: true, ical.PropStatus: true,
-	ical.PropTransparency: true, "SEQUENCE": true, ical.PropPriority: true,
-	ical.PropClass: true, ical.PropURL: true, "CONFERENCE": true, ical.PropGeo: true,
-	ical.PropCategories: true, ical.PropExceptionDates: true,
-	ical.PropRecurrenceDates: true, ical.PropRecurrenceID: true,
-	ical.PropDateTimeStamp: true, ical.PropCreated: true, ical.PropLastModified: true,
-	ical.PropAttach: true, ical.PropComment: true, ical.PropContact: true,
-	ical.PropResources: true, ical.PropRelatedTo: true,
-	ical.PropAttendee: true, ical.PropOrganizer: true,
-}
-
-// handledTodoProps is the set of property names explicitly parsed by todoFromVTodo.
-var handledTodoProps = map[string]bool{
-	ical.PropUID: true, ical.PropSummary: true, ical.PropDescription: true,
-	ical.PropLocation: true, ical.PropDateTimeStart: true, ical.PropDue: true,
-	ical.PropDuration: true, ical.PropCompleted: true, ical.PropPercentComplete: true,
-	ical.PropRecurrenceRule: true, ical.PropStatus: true,
-	"SEQUENCE": true, ical.PropPriority: true,
-	ical.PropClass: true, ical.PropURL: true, ical.PropGeo: true,
-	ical.PropCategories: true, ical.PropExceptionDates: true,
-	ical.PropRecurrenceDates: true, ical.PropRecurrenceID: true,
-	ical.PropDateTimeStamp: true, ical.PropCreated: true, ical.PropLastModified: true,
-	ical.PropAttach: true, ical.PropComment: true, ical.PropContact: true,
-	ical.PropResources: true, ical.PropRelatedTo: true,
-	ical.PropAttendee: true, ical.PropOrganizer: true,
-}
-
-// handledJournalProps is the set of property names explicitly parsed by journalFromVJournal.
-var handledJournalProps = map[string]bool{
-	ical.PropUID: true, ical.PropSummary: true, ical.PropDescription: true,
-	ical.PropDateTimeStart: true, ical.PropRecurrenceRule: true, ical.PropStatus: true,
-	"SEQUENCE": true, ical.PropClass: true, ical.PropURL: true,
-	ical.PropCategories: true, ical.PropExceptionDates: true,
-	ical.PropRecurrenceDates: true, ical.PropRecurrenceID: true,
-	ical.PropDateTimeStamp: true, ical.PropCreated: true, ical.PropLastModified: true,
-	ical.PropAttach: true, ical.PropComment: true, ical.PropContact: true,
-	ical.PropRelatedTo: true,
-	ical.PropAttendee:  true, ical.PropOrganizer: true,
 }
 
 // extractXPropertiesWithSet collects properties not in the handled set.
